refactor event edit into component and embed in event page
